feat: add netlify function for express backend
