+++ RenovatePro/netlify/functions/api.ts
@@ -1,0 +1,6 @@
+import serverless from "serverless-http";
+import { createApp } from "../../server/app";
+
+const app = await createApp();
+
+export const handler = serverless(app);
