+++ netlify.toml
@@ -1,0 +1,5 @@
+[build]
+  base = "RenovatePro"
+  command = "npm run build"
+  publish = "dist/public"
+  functions = "netlify/functions"
